+++ SCEM Web/src/css/userGroup.css
@@ -1,180 +1,68 @@
-#userGroup-body {
+#userGroups-body {
     text-align: center;
     font-family: Arial, Helvetica, sans-serif;
 }
 
-#userGroup-header {
+#userGroups-header {
     background-color: #133da5;
-    padding-bottom: 4em;
-    padding-top: 2em;
+    padding: 2em 0 0 0;
     border-top-left-radius: 2em;
     border-top-right-radius: 2em;
     margin: 2em 30% 2em;
 }
 
-#userGroup-titleText {
+#userGroups-titleText {
     color: white;
     display: inline-block;
 }
 
-#userGroup-backButton {
+#userGroups-backButton {
     position: relative;
     display: inline-block;
     right: 4em;
+    padding-bottom: 4em;
     bottom: 0.1em;
     filter: invert(100%);
 }
 
-#userGroup-body {
-    width: 400px;
-    margin: 0 auto;
-    padding: 20px;
+#userGroups-addButtonDiv {
+    position: relative;
+    display: inline-block;
+    height: 5px;
+    width: 5px;
+    left: 20%;
+    top: 0.7em;
 }
 
-#userGroup-inputBox {
-    margin-bottom: 20px;
+#userGroups-addButton {
+    height: 40px;
+    width: 40px;
 }
 
-#userGroup-inputBox input[type="text"] {
-    width: 100%;
-    padding: 15px;
-    font-size: 16px;
+#userGroups-table {
+    margin-left: auto;
+    margin-right: auto;
+    counter-reset: rowNumber;
 }
 
-#userGroup-permissionsHeader {
-    font-size: 18px;
-    margin-bottom: 10px;
-    text-align: left;
+.userGroups-tableElement {
+    text-align: center;
+    border: 1px solid black;
+    height: 2.5em;
 }
 
-#userGroup-checkboxes {
-    display: flex;
-    flex-direction: column;
-}
-
-#userGroup-checkboxes label {
-    display: flex;
-    align-items: center;
-    font-size: 16px;
-    margin-bottom: 10px;
-}
-
-#userGroup-checkboxes input[type="checkbox"] {
-    margin-right: 10px;
-    width: 20px;
-    height: 20px;
-}
-
-#userGroup-addUserSection {
-    display: flex;
-    align-items: center;
-    margin-top: 10px;
-}
-
-#userGroup-addUserSection input[type="text"] {
-    flex: 1;
-    padding: 15px; /* Increase padding for height */
-    font-size: 18px; /* Increase font size */
-    margin-right: 10px;
-}
-
-#userGroup-addButton {
-    background-color: blue;
-    color: white;
-    border: none;
-    border-radius: 5px;
-    padding: 15px 24px; /* Increase padding for height and width */
-    font-size: 18px; /* Increase font size */
-    cursor: pointer;
-}
-
-#userGroup-table {
-    width: 100%;
-    border-collapse: collapse;
-}
-
-#userGroup-tableHead {
+.userGroups-tableHeading {
+    border: 1px solid black;
+    height: 2.5em;
     background-color: #5982d3;
 }
 
-.userGroup-tableHeading {
-    padding: 10px;
-    text-align: center;
-    border: 1px solid #ccc; /* Add grey border to table header cells */
+#userGroups-table tr:not(:first-child) {
+    counter-increment: rowNumber;
 }
 
-#userGroup-tableBody td {
-    padding: 10px;
-    border: 1px solid #ccc; /* Add grey border to table cells */
-}
-
-#userGroup-table button {
-    background-color: transparent;
-    border: none;
-    cursor: pointer;
-    font-size: 14px;
-    color: #ff0000;
-}
-
-#userGroup-table button:hover {
-    color: #ff3333;
-}
-
-.remove-button {
-    background-color: transparent;
-    color: red;
-    border: none;
-    padding: 12px 20px; /* Increase padding to make the button thicker */
-    cursor: pointer;
-    font-weight: bold;
-    font-size: 25px; /* Adjust font size as needed */
-    height: 50px; /* Set a specific height for the button */
-}
-
-#userGroup-saveButton {
-    background-color: blue;
-    color: white;
-    border: none;
-    border-radius: 5px;
-    padding: 15px 170px; /* Increase padding for height and width */
-    font-size: 18px; /* Increase font size */
-    cursor: pointer;
-}
-
-.checkbox {
-    display: none;
-}
-
-.checkbox-label {
-    position: relative;
-    cursor: pointer;
-    padding-left: 25px;
-}
-
-.checkbox-label::before {
-    content: "";
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 18px;
-    height: 18px;
-    border: 2px solid blue;
-    border-radius: 4px;
-}
-
-.checkbox-label::after {
-    content: "";
-    position: absolute;
-    top: 3px;
-    left: 5px;
-    width: 7px;
-    height: 12px;
-    border: solid blue;
-    border-width: 0 2px 2px 0;
-    transform: rotate(45deg);
-    opacity: 0;
-}
-
-.checkbox:checked + .checkbox-label::after {
-    opacity: 1;
+#userGroups-table tr .userGroups-tableElement:first-child::before {
+    content: counter(rowNumber);
+    margin-left: 0.5em;
+    margin-right: 0.5em;
 }
